bot/pyproject.toml
[project]
name = "pereobuyka"
version = "0.1.0"
description = "Telegram-бот-консультант для записи в сервис шиномонтажа."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "aiogram>=3.0.0",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.0",
]

[project.scripts]
pereobuyka = "pereobuyka.main:main"

[build-system]
requires = ["uv_build>=0.8.0,<0.9.0"]
build-backend = "uv_build"

[tool.uv]
package = true

[dependency-groups]
dev = [
  "mypy>=1.20.1",
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "ruff>=0.5.0",
  "ty>=0.0.1a16",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

# Стиль: см. .agents/skills/python-code-style — ruff как единый линтер/форматтер
[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "SIM"]
# Длинные строки — зона форматтера; многострочные литералы (промпты) не дробим принудительно
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.ty.environment]
python-version = "3.12"
